# docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Harmony API
info:
  description: |
    This is the API documentation for Harmony.
  version: 1.0.0
  title: Harmony API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  "/api/v1/reports":
    get:
      tags:
        - Reports General Operations
      summary: Get all reports
      operationId: getAllReports
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: userid
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Search successful
          content:
            "Response Body":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Reports"
        '403':
          description: Not enough permissions
    post:
      tags:
        - Reports General Operations
      summary: Report an user's review
      operationId: postReport
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReportRequestHelper"
        required: true
      responses:
        '201':
          description: Item created
          content:
            "Response Body":
              schema:
                "$ref": "#/components/schemas/ReportResponseHelper"
        '400':
          description: Some parameter does not have a valid value
        '409':
          description: Item already exists
  "/api/v1/reports/{id}":
    delete:
      tags:
        - Reports Specific Operations
      summary: Deletes a report.
      operationId: deleteReport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UseridBodyHelper"
        required: true
      responses:
        '204':
          description: Item destroyed
        '400':
          description: Some parameter does not have a valid value
        '403':
          description: Not enough permissions
        '404':
          description: Item doesn't exists
components:
  schemas:
...
    UseridBodyHelper:
      type: object
      properties:
        userid:
          type: integer
          format: int32
...
    ReportRequestHelper:
      type: object
      properties:
        useridreporter:
          type: integer
          format: int32
        useridreported:
          type: integer
          format: int32
        reviewid:
          type: integer
          format: int32
        reason:
          type: string
    ReportResponseHelper:
      type: object
      properties:
        reportid:
          type: integer
          format: int32
        useridreporter:
          type: integer
          format: int32
        useridreported:
          type: integer
          format: int32
        reviewid:
          type: integer
          format: int32
        reason:
          type: string
...
    Reports:
      type: object
      properties:
        reportid:
          type: integer
          format: int32
        useridreporter:
          type: integer
          format: int32
        useridreported:
          type: integer
          format: int32
        reviewid:
          type: integer
          format: int32
        reason:
          type: string
...